Move get_puppet_services_that_run_here(_for_nerve) to its own file, puppet_service_tools.py; clean up unused get_classic_service_information_for_nerve
